backend/pyproject.toml
[tool.poetry]
name = "django-phone-auth"
version = "0.0.1"
description = ""
authors = ["Nikita <vekshinnikita>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
